+++ packages/upload-web/src/index.ts
@@ -1,3 +1,5 @@
 import {Builder} from "./builder";
 export default Builder
 export { Builder as WebUploader}
+
+export type * from "./builder"