+++ packages/ethereum-web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @irys/web-upload-ethereum
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @irys/web-upload@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes